+++ app/src/main/java/org/simple/clinic/newentry/country/InputFieldsProvider.kt
@@ -1,0 +1,7 @@
+package org.simple.clinic.newentry.country
+
+import org.simple.clinic.newentry.form.InputField
+
+interface InputFieldsProvider {
+  fun provide(): List<InputField<*>>
+}